First Implementation of Actor Data and Traffic Light Fallback Sensors
The fallback sensors are implemented in a quick and straightforward manner to grant ease of use.
The sensors are implemented as ActorComponents that are spawned when the game mode detects the ego vehicle tags.
From thereafter the sensors collects per frame data according to the specifications.

TODO:
TrafficLight sensor name filter is not working
ActorData sensor detects all actors in scene tracked by the registry, further filtering is necessary to avoid including non-mesh actors like sensors